break out separate factory and field modules
