+++ objectfactory/__init__.py
@@ -2,12 +2,8 @@
 objectfactory
 """
 
-from .serializable import (
-    Field,
-    Serializable,
-    Factory,
-    Nested,
-    List
-)
+from .serializable import Serializable, Field
+from .factory import Factory
+from .field import Nested, List
 
 __version__ = '0.0.1'
